+++ src/api/objects/toot.ts
@@ -12,7 +12,7 @@ import MastoApi from "../api";
 import MastodonServer from "../mastodon_server";
 import Scorer from "../../scorer/scorer";
 import UserData from "../user_data";
-import { ageInHours, ageInSeconds, ageString, timelineCutoffAt, toISOFormat } from "../../helpers/time_helpers";
+import { ageInHours, ageInMinutes, ageInSeconds, ageString, timelineCutoffAt, toISOFormat } from "../../helpers/time_helpers";
 import { batchMap, groupBy, sortObjsByProps, sumArray, uniquify, uniquifyByProp } from "../../helpers/collection_helpers";
 import { Config } from "../../config";
 import { FOREIGN_SCRIPTS, LANGUAGE_CODES, detectLanguage } from "../../helpers/language_helper";
@@ -631,7 +631,7 @@ export default class Toot implements TootObj {
 
     // Returns true if the toot should be re-completed
     private shouldComplete(): boolean {
-        return !this.completedAt || (ageInSeconds(this.completedAt) > (Config.staleDataTrendingSeconds || 3600));
+        return !this.completedAt || (ageInMinutes(this.completedAt) > Config.staleDataTrendingMinutes);
     }
 
     ///////////////////////////////