+++ src/helpers/collection_helpers.ts
@@ -132,7 +132,7 @@ export function filterWithLog<T>(
 // the minimum ID in a set of toots being wildly out of step with the rest of the IDs.
 // If that happens trying to use the min ID as the maxId param for a fetch will fail (no results).
 // This is an unfixable server side problem that we work around in TheAlgorithm.maybeFetchMoreData()
-export function findMinId(array: MastodonObjWithID[]): string | undefined{
+export function findMinMaxId(array: MastodonObjWithID[]): [string, string] | undefined {
     if (array.length == 0) return undefined;
     const idVals = array.map(e => e.id);
     const isNumberArray = idVals.every(isNumber);
@@ -149,7 +149,7 @@ export function findMinId(array: MastodonObjWithID[]): string | undefined{
         }
     });
 
-    return sortedIDs[0].toString();
+    return [sortedIDs[0].toString(), sortedIDs.slice(-1)[0].toString()];
 };